Add stat for PRs merged with no retest required
This update allows us to measure the number of PRs that are merged
without requiring any retests over a period of time.

It also adds a badge to README.md that will be updated by the existing
automation with the number of PRs merged without retests in the past 7
days.

Signed-off-by: Brian Carey <bcarey@redhat.com>