Update time to merge badge from days to hours
Thanks to improvements over the last few months the average time to
merge is no longer mulitple days but now can be measured in hours.

This change only updates the badge to display the time to merge in hours
rather than days.

The underlying data has not been changed as to keep the historical
results and graphs consistent.

Signed-off-by: Brian Carey <bcarey@redhat.com>