retests: Remove check for /test from retest count
There are occasions where contributors want to test some lanes multiple
times against a PR to ensure that the PR does not introduce an
instability.

In these cases, the contributors use /test to trigger these lanes. This
muddies the PRs merged with zero retest metric as these are not retests
but are counted as such. It is also used to trigger test lanes for
contributors who are not yet kubevirt members and have permissions to
run CI.

For example in these two recent PRs:
https://github.com/kubevirt/kubevirt/pull/14880
https://github.com/kubevirt/kubevirt/pull/14832

Remove the check for /test as this causes PRs with 0 retest to be missed

Signed-off-by: Brian Carey <bcarey@redhat.com>